+++ geocube/api/core.py
@@ -2,6 +2,8 @@
 """
 GeoCube client core functionality
 """
+import numpy
+
 from geocube.geo_utils.geobox import GeoBoxMaker
 from geocube.vector_to_cube import VectorToCube
 
@@ -15,7 +17,7 @@ def make_geocube(
     align=None,
     geom=None,
     like=None,
-    fill=None,
+    fill=numpy.nan,
     group_by=None,
     interpolate_na_method=None,
     categorical_enums=None,
@@ -59,7 +61,7 @@ def make_geocube(
             gcds = make_geocube(vector_data='my_vector.geopackage', like=other_gcds)
 
     fill: float, optional
-        The value to fill in the grid with for nodata. Default is -9999.0.
+        The value to fill in the grid with for nodata. Default is NaN.
     group_by: str, optional
         When specified, perform basic combining/reducing of the data on this column.
     interpolate_na_method:  {‘linear’, ‘nearest’, ‘cubic’}, optional