+++ geocube/cli/commands/make_geocube.py
@@ -4,6 +4,7 @@ The CLI interface to the load module for GeoCube
 """
 
 import click
+import numpy
 import xarray
 
 from geocube.api import core
@@ -27,7 +28,8 @@ from geocube.api import core
     "-f",
     "--fill",
     type=float,
-    help="The value to fill in the grid with for nodata. Default is -9999.0.",
+    help="The value to fill in the grid with for nodata. Default is NaN.",
+    default=numpy.nan,
     required=False,
 )
 @click.option(